fix conversion warnings
git-svn-id: http://localhost:9111/private/safir_open/safir_sdk_core/trunk@1025 9dd7cf02-b8c9-4ff8-a4a5-e7fda3bdd973

+++ src/dope/dope_main.ss/src/dope_main/FilePersistor.cpp
@@ -298,7 +298,7 @@ FilePersistor::RestoreBinary(const boost::filesystem::path & path) const
     while (file.good())
     {
         file.read(&bin[0] + numBytesRead,4096);
-        numBytesRead += file.gcount();
+        numBytesRead += static_cast<size_t>(file.gcount());
     }
     if(fileSize != numBytesRead)
     {
@@ -335,7 +335,7 @@ FilePersistor::RestoreXml(const boost::filesystem::path & path) const
     while (file.good())
     {
         file.read(&xml[0] + numBytesRead,4096);
-        numBytesRead += file.gcount();
+        numBytesRead += static_cast<size_t>(file.gcount());
     }
     if(fileSize < numBytesRead)
     {
